refactor(wallet-mobile-modal): modularize MobileDropdown into view components and improve UI
- Broke down MobileDropdown into WalletView, TransferView, FundView, HistoryView, and SettingsView for better maintainability and clarity.
- Moved logic and state into respective subcomponents, making MobileDropdown a clean orchestrator.
- Ensured all mobile modal views match desktop modal functionality, validation, and error handling.
- Improved UI/UX consistency, including button placements and hover states.
- Removed unused code and cleaned up imports for a more maintainable codebase.
