+++ app/components/wallet-mobile-modal/index.ts
@@ -1,0 +1,5 @@
+export { WalletView } from "./WalletView";
+export { TransferView } from "./TransferView";
+export { FundView } from "./FundView";
+export { HistoryView } from "./HistoryView";
+export { SettingsView } from "./SettingsView";
